+++ src/greentest/test__core.py
@@ -1,4 +1,5 @@
-# pylint:disable=no-member
+
+from __future__ import absolute_import, print_function, division
 import sys
 import unittest
 import greentest
@@ -6,13 +7,14 @@ import greentest
 from gevent import core
 
 
+
 class TestCore(unittest.TestCase):
 
     def test_get_version(self):
-        version = core.get_version()
+        version = core.get_version() # pylint: disable=no-member
         self.assertIsInstance(version, str)
         self.assertTrue(version)
-        header_version = core.get_header_version()
+        header_version = core.get_header_version() # pylint: disable=no-member
         self.assertIsInstance(header_version, str)
         self.assertTrue(header_version)
         self.assertEqual(version, header_version)
@@ -20,8 +22,18 @@ class TestCore(unittest.TestCase):
 
 class TestWatchers(unittest.TestCase):
 
-    def makeOne(self):
-        return core.loop()
+    def _makeOne(self):
+        return core.loop() # pylint:disable=no-member
+
+    def destroyOne(self, loop):
+        loop.destroy()
+
+    def setUp(self):
+        self.loop = self._makeOne()
+
+    def tearDown(self):
+        self.destroyOne(self.loop)
+        del self.loop
 
     def test_io(self):
         if sys.platform == 'win32':
@@ -31,47 +43,62 @@ class TestWatchers(unittest.TestCase):
         else:
             Error = ValueError
             win32 = False
+
         with self.assertRaises(Error):
-            self.makeOne().io(-1, 1)
+            self.loop.io(-1, 1)
+
         if hasattr(core, 'TIMER'):
             # libev
             with self.assertRaises(ValueError):
-                self.makeOne().io(1, core.TIMER)
+                self.loop.io(1, core.TIMER) # pylint:disable=no-member
 
         # Test we can set events and io before it's started
         if not win32:
             # We can't do this with arbitrary FDs on windows;
             # see libev_vfd.h
-            io = self.makeOne().io(1, core.READ)
+            io = self.loop.io(1, core.READ) # pylint:disable=no-member
             io.fd = 2
             self.assertEqual(io.fd, 2)
-            io.events = core.WRITE
+            io.events = core.WRITE # pylint:disable=no-member
             if not hasattr(core, 'libuv'):
                 # libev
+                # pylint:disable=no-member
                 self.assertEqual(core._events_to_str(io.events), 'WRITE|_IOFDSET')
             else:
-                self.assertEqual(core._events_to_str(io.events), 'WRITE')
+
+                self.assertEqual(core._events_to_str(io.events), # pylint:disable=no-member
+                                 'WRITE')
             io.start(lambda: None)
             io.close()
 
+
     def test_timer_constructor(self):
         with self.assertRaises(ValueError):
-            self.makeOne().timer(1, -1)
+            self.loop.timer(1, -1)
 
     def test_signal_constructor(self):
         with self.assertRaises(ValueError):
-            self.makeOne().signal(1000)
+            self.loop.signal(1000)
 
 class TestWatchersDefault(TestWatchers):
 
-    def makeOne(self):
-        return core.loop(default=True)
+    def _makeOne(self):
+        return core.loop(default=True) # pylint:disable=no-member
 
-@greentest.skipOnLibuvOnPyPyOnWin("This crashes with PyPy 5.10.0, only on Windows. "
-                                  "See https://ci.appveyor.com/project/denik/gevent/build/1.0.1380/job/lrlvid6mkjtyrhn5#L1103")
+    def destroyOne(self, loop):
+        return
+
+# XXX: The crash may be fixed? The hang showed up after the crash was
+# reproduced and fixed on linux and OS X.
+@greentest.skipOnLibuvOnWin(
+    "This crashes with PyPy 5.10.0, only on Windows. "
+    "See https://ci.appveyor.com/project/denik/gevent/build/1.0.1380/job/lrlvid6mkjtyrhn5#L1103 "
+    "It has also timed out, but only on Appveyor CPython 3.6; local CPython 3.6 does not. "
+    "See https://ci.appveyor.com/project/denik/gevent/build/1.0.1414/job/yn7yi8b53vtqs8lw#L1523")
 class TestWatchersDefaultDestroyed(TestWatchers):
 
-    def makeOne(self):
+    def _makeOne(self):
+        # pylint: disable=no-member
         l = core.loop(default=True)
         l.destroy()
         del l
@@ -81,6 +108,7 @@ class TestWatchersDefaultDestroyed(TestWatchers):
 class TestLibev(unittest.TestCase):
 
     def test_flags_conversion(self):
+        # pylint: disable=no-member
         if sys.platform != 'win32':
             self.assertEqual(core.loop(2, default=False).backend_int, 2)
         self.assertEqual(core.loop('select', default=False).backend, 'select')
@@ -94,11 +122,14 @@ class TestLibev(unittest.TestCase):
 class TestEvents(unittest.TestCase):
 
     def test_events_conversion(self):
-        self.assertEqual(core._events_to_str(core.READ | core.WRITE), 'READ|WRITE')
+        self.assertEqual(core._events_to_str(core.READ | core.WRITE), # pylint: disable=no-member
+                         'READ|WRITE')
 
     def test_EVENTS(self):
-        self.assertEqual(str(core.EVENTS), 'gevent.core.EVENTS')
-        self.assertEqual(repr(core.EVENTS), 'gevent.core.EVENTS')
+        self.assertEqual(str(core.EVENTS), # pylint: disable=no-member
+                         'gevent.core.EVENTS')
+        self.assertEqual(repr(core.EVENTS), # pylint: disable=no-member
+                         'gevent.core.EVENTS')
 
 if __name__ == '__main__':
     greentest.main()